+++ CitationAddViewController.h
@@ -13,7 +13,7 @@
 
 @interface CitationAddViewController : UIViewController {
 	IBOutlet UIPickerView*							citationTypePicker;
-	int												column0Selection, column1Selection;
+	NSInteger										column0Selection, column1Selection;
 	IBOutlet CitationDetailTableViewController*		citationDetailController;
 	NSString*										paperName;
 	TermPaperModel*									mTermPaper;
@@ -22,7 +22,7 @@
 - (IBAction)handleAddCitation:(id)sender;
 
 @property (nonatomic, strong) IBOutlet UIPickerView *citationTypePicker;
-@property (nonatomic) int column0Selection, column1Selection;
+@property (nonatomic) NSInteger column0Selection, column1Selection;
 @property (nonatomic, strong) IBOutlet CitationDetailTableViewController *citationDetailController;
 @property (nonatomic, strong) NSString *paperName;
 @property (nonatomic, strong) TermPaperModel *mTermPaper;
